clase6/practica/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
pandas
google-cloud-storage
pydantic
gunicorn